+++ lib/features/store/screens/store_items_tings/checkout/widgets/cart_items.dart
@@ -171,7 +171,7 @@ class CCartItems extends StatelessWidget {
                                   final thisCartItem = cartController
                                       .convertInvToCartItem(
                                         invItem,
-                                        int.parse(value),
+                                        double.parse(value),
                                       );
 
                                   cartController.addSingleItemToCart(
